backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
pwdlib[argon2]
python-multipart
pydantic-settings
email-validator
boto3
openai>=2.0.0
requests
